requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
mysqlclient==2.2.7
peewee==3.17.9
pymssql==2.3.2
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
